Generalise the struct comparison
Both structs have char * as their first elements.
Pointers to the structs are pointers to each of their first elements.
(https://beej.us/guide/bgc/html/split/structs-ii-more-fun-with-structs.html#fake-oop)
So, they would be char ** (not char *, as they're pointers to char *).
We then take the objects they point to, to get char *, so that we can
compare them with strcmp.